Update envOrError to optionally avoid type conversion
In this change, a noConvert option is added to the envOrError function. This option, when set to true, will return the environment variable's value as is, without converting its type. This change is made to offer more flexibility in handling environment variables, especially when the type conversion is not necessary or desirable.